requirements.txt
streamlit
langchain-openai
python-dotenv
Pillow
langchain
